+++ joint_trajectory_controller/include/joint_trajectory_controller/hardware_interface_adapter.h
@@ -1,0 +1,200 @@
+///////////////////////////////////////////////////////////////////////////////
+// Copyright (C) 2013, PAL Robotics S.L.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//   * Redistributions of source code must retain the above copyright notice,
+//     this list of conditions and the following disclaimer.
+//   * Redistributions in binary form must reproduce the above copyright
+//     notice, this list of conditions and the following disclaimer in the
+//     documentation and/or other materials provided with the distribution.
+//   * Neither the name of hiDOF, Inc. nor the names of its
+//     contributors may be used to endorse or promote products derived from
+//     this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+//////////////////////////////////////////////////////////////////////////////
+
+/// \author Adolfo Rodriguez Tsouroukdissian
+
+#ifndef JOINT_TRAJECTORY_CONTROLLER_HARDWARE_INTERFACE_ADAPTER_H
+#define JOINT_TRAJECTORY_CONTROLLER_HARDWARE_INTERFACE_ADAPTER_H
+
+#include <cassert>
+#include <string>
+#include <vector>
+
+#include <boost/shared_ptr.hpp>
+
+#include <ros/node_handle.h>
+#include <ros/time.h>
+
+#include <control_toolbox/pid.h>
+#include <hardware_interface/joint_command_interface.h>
+
+/**
+ * \brief Helper class to simplify integrating the JointTrajectoryController with different hardware interfaces.
+ *
+ * The JointTrajectoryController outputs position, velocity, command triplets, while the more common hardware interfaces
+ * it is supposed to work with accept either position or effort commands.
+ *
+ * Use one of the avaialble template specializations of this class (or create your own) to adapt the
+ * JointTrajectoryController to a specidfic hardware interface.
+ */
+template <class HardwareInterface, class State>
+class HardwareInterfaceAdapter
+{
+public:
+  bool init(const std::vector<hardware_interface::JointHandle>& joint_handles, ros::NodeHandle& controller_nh)
+  {
+    return false;
+  }
+
+  void starting(const ros::Time& time) {}
+  void stopping(const ros::Time& time) {}
+
+  void update(const ros::Time&     time,
+              const ros::Duration& period,
+              const State&         desired_state,
+              const State&         state_error,
+              std::vector<double>& command) {}
+};
+
+/**
+ * \brief Adapter for a position-controlled hardware interface. Forwards desired positions as commands.
+ */
+template <class State>
+class HardwareInterfaceAdapter<hardware_interface::PositionJointInterface, State>
+{
+public:
+  bool init(const std::vector<hardware_interface::JointHandle>& joint_handles, ros::NodeHandle& controller_nh)
+  {
+    return true;
+  }
+
+  void starting(const ros::Time& time) {}
+  void stopping(const ros::Time& time) {}
+
+  void update(const ros::Time&     /*time*/,
+              const ros::Duration& /*period*/,
+              const State&         desired_state,
+              const State&         /*state_error*/,
+              std::vector<double>& command)
+  {
+    // Precondition
+    assert(desired_state.position.size() == command.size());
+
+    // Forward desired position to command
+    for (unsigned int i = 0; i < command.size(); ++i) {command[i] = desired_state.position[i];}
+  }
+};
+
+/**
+ * \brief Adapter for an effort-controlled hardware interface. Maps position and velocity errors to effort commands
+ * through a position PID loop.
+ *
+ * The following is an example configuration of a controller that uses this adapter. Notice the \p gains entry:
+ * \code
+ * head_controller:
+ *   type: "position_controllers/JointTrajectoryController"
+ *   joints:
+ *     - head_1_joint
+ *     - head_2_joint
+ *
+ *   constraints:
+ *     goal_time: 0.6
+ *     stopped_velocity_tolerance: 0.02
+ *     head_1_joint:
+ *       trajectory: 0.05
+ *       goal: 0.02
+ *     head_2_joint:
+ *       goal: 0.01
+ *
+ *   gains:
+ *     head_1_joint: {p: 200, d: 1, i: 5, i_clamp: 1}
+ *     head_2_joint: {p: 200, d: 1, i: 5, i_clamp: 1}
+ * \endcode
+ */
+template <class State>
+class HardwareInterfaceAdapter<hardware_interface::EffortJointInterface, State>
+{
+public:
+  HardwareInterfaceAdapter() : joint_handles_ptr_(0) {}
+
+  bool init(const std::vector<hardware_interface::JointHandle>& joint_handles, ros::NodeHandle& controller_nh)
+  {
+    // Store pointer to joint handles
+    joint_handles_ptr_ = &joint_handles;
+
+    // Initialize PIDs
+    pids_.resize(joint_handles.size());
+    for (unsigned int i = 0; i < pids_.size(); ++i)
+    {
+      // Node handle to PID gains
+      ros::NodeHandle joint_nh(controller_nh, std::string("gains/") + joint_handles[i].getName());
+
+      // Init PID gains from ROS parameter server
+      pids_[i].reset(new control_toolbox::Pid());
+      if (!pids_[i]->init(joint_nh))
+      {
+        ROS_WARN_STREAM("Failed to initialize PID gains from ROS parameter server.");
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  void starting(const ros::Time& time)
+  {
+    if (!joint_handles_ptr_) {return;}
+
+    // Reset PIDs, zero effort commands
+    for (unsigned int i = 0; i < pids_.size(); ++i)
+    {
+      pids_[i]->reset();
+      (*joint_handles_ptr_)[i].setCommand(0.0);
+    }
+  }
+
+  void stopping(const ros::Time& time) {}
+
+  void update(const ros::Time&     /*time*/,
+              const ros::Duration& period,
+              const State&         /*desired_state*/,
+              const State&         state_error,
+              std::vector<double>& command)
+  {
+    const unsigned int n_joints = pids_.size();
+
+    // Preconditions
+    if (!joint_handles_ptr_) {return;}
+    assert(n_joints == state_error.position.size());
+    assert(n_joints == state_error.velocity.size());
+    assert(n_joints == command.size());
+
+    // Update PIDs
+    for (unsigned int i = 0; i < n_joints; ++i)
+    {
+      command[i] = pids_[i]->computeCommand(state_error.position[i], state_error.velocity[i], period);
+    }
+  }
+
+private:
+  typedef boost::shared_ptr<control_toolbox::Pid> PidPtr;
+  std::vector<PidPtr> pids_;
+
+  std::vector<hardware_interface::JointHandle>* joint_handles_ptr_;
+};
+
+#endif // header guard
