+++ joint_trajectory_controller/ros_control_plugins.xml
@@ -1,8 +1,20 @@
 <library path="lib/libjoint_trajectory_controller">
 
-  <class name="position_controllers/JointTrajectoryController" type="joint_trajectory_controller::JointTrajectoryController" base_class_type="controller_interface::ControllerBase">
+  <class name="position_controllers/JointTrajectoryController"
+         type="position_controllers::JointTrajectoryController"
+         base_class_type="controller_interface::ControllerBase">
     <description>
       The JointTrajectoryController executes joint-space trajectories on a set of joints.
+      This variant represents trajectory segments as quintic splines and sends commands to a position interface.
+    </description>
+  </class>
+
+  <class name="effort_controllers/JointTrajectoryController"
+         type="effort_controllers::JointTrajectoryController"
+         base_class_type="controller_interface::ControllerBase">
+    <description>
+      The JointTrajectoryController executes joint-space trajectories on a set of joints.
+      This variant represents trajectory segments as quintic splines and sends commands to an effort interface.
     </description>
   </class>
